requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.18
aiofiles==24.1.0
python-dotenv==1.0.1

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
